Roll back platform to avoid Job scheduling issue
The 2022-12 M1 build from Platform may have a regression that is causing
CDT build to fail more often due to deadlock in the indexer.

This commit uses 2022-09 R build of Platform to avoid the problem.

Note that this won't resolve the issue in Platform, but will allow
CDT to be built.

Fixes #81