Fixes launchbar icons looking blurry on hidpi screens
The launchbar way of compositing images for the buttons was preventing
the use of higher quality icons on hidpi dislays, making the launchbar
buttons look blurry or pixelated compared to other toolbar buttons.

Instead we can get a nicer result with a custom widget on which we can
directly draw a button border plus the hidpi version of the icon.

Also included is an attempt at a 2x hammer build button icon, which
should look a bit nicer on hidpi screens.