Correct descriptorTypeId for default launch type
This is mostly a no-op change, but fixes the error
about incorrect reference.

The descriptorUI extension used an ID that never existed,
and the code in ConfigSelector just fell back on
DefaultDescriptorLabelProvider it none were found, so
the user visible change here is nothing.

Fixup for 8f8c2e02b679852087c419a5f8c9531315934b01