Docker: pull image automatically (#96)
* Bug 566810: Pull Docker image if necessary

Change-Id: I948e3a273ae598efe563b150922a7bc4b696d05c
Signed-off-by: Moritz 'Morty' Strübe <moritz.struebe@mathema.de>

* docker.launcher: Add tests

* Docker: Fix NPE

This NPE was triggered when the docker image was deleted in the
background.

Change-Id: I7fe8f9204a2f60a5d2396bbc2ec67c7ae820a60a
Signed-off-by: Moritz 'Morty' Strübe <moritz.struebe@mathema.de>

* Bug 566810 Allow setting any Docker image

Allow setting any docker image as string in the input field. There is no
verification yet. If the docker image exists, it is set accordingly.

Change-Id: Iec21798cf79206151067d1e2f3a7ec933f1d0b7d
Signed-off-by: Moritz 'Morty' Strübe <moritz.struebe@mathema.de>

* Bug 566810 Remove volumes immediately

As it is now possible to insert any string, the volumes added by the
image, should be removed immediately, too.

Change-Id: I6c843a2b1e551372e0dff22185d6152faecd72f3

Signed-off-by: Moritz 'Morty' Strübe <moritz.struebe@mathema.de>