Simplify API document
Removed the manually maintained table of contents, instead use
the contents feature in GitHub to navigate. The spans have
all been removed because they were not working anyway, at least
on github.

This change was inspired by PR #336 where a contributor was updating
this document and I didn't think it was a great idea to send it back
to them to add the TOC entry.