Lock snakeyaml to expected 1.x version
Snakeyml recentlyish came out with 2.0 version and more recently
TM4E started using the 2.0 version. As TM4E snapshots are in our
target platform we started failing as we started wiring to the newer
version due to a missing underbound on our dependency.

A separate task of updating to recent snakeyaml will be done as part
of #387