Remove assumption that strings are always intern
The pre-existing code assumed that string (qName) passed in
to start and end element was always interned and compared
those strings to constants using ==. Instead rewrite
code using switch statements on the strings.

Also includes small change to the exception handling so that
code analysis can correctly where exceptions are always thrown.
i.e. instead of hiding a throw in an always throwing method,
return the exception and throw it at the call site.