Improve memory search results page
The Memory Browser and Memory views will jump to the memory
address after a user selects a value from the Memory Search
Result list. If the views belong to the same tab group as
the Search view, then it might not be very intuitive for
the user to switch back to one of the Memory views and see
the effect of the selection.

The current commit adds on the the double click listener
(from MemorySearchResultsPage.java) the behavior to refocus
on the Memory Browser/ Memory view.

Resolves: #515