Run only the code cleanliness on Jenkins CI
I recently split code cleanliness into two parts, and since in Jenkins
the baseline-compare-and-replace and api-baseline-check run on the main
build we don't need to run them on the cleanliness pass too. Therefore
run the "only" script in the code formatting checks stage to speed up
build and avoid duplicated work.