Allocate much more ram to the Pod on Jenkins
I think with some of the recent changes (Tycho 4, API baselines,
maybe even new dependency on Platform M3) it may be that our memory
requirements have gone up substantially for the build.

Also, with Sonar in the works that also requires more memory.

Therefore see if the EF's JIPP infra will allocate 10G of ram to
our build.

This should fix all the "Killed" messages randomly in the CDT builds.
https://wiki.eclipse.org/Jenkins#What_is_killing_my_build.3F_I.27m_using_custom_containers.21