Fix disassembly view when view is reset
The view can be reset in a couple of ways, by changing debug
context or manually refreshing. With this change in Platform
https://github.com/eclipse-platform/eclipse.platform.ui/pull/963
the CDT violation of the API aronud resetting state of the
document was exposed.

Therefore the solution is to instead of trying to reset the
state of the existing document, create a new one when the
view is reset.

Fixes #603