Ensure DSF GDB preference pages and UI elements are available in Core Build
When using Core Build, the GdbLaunchDelegate is not used to create
the launch. Therefore in the core launch we need to enable the
activity to show all the DSF GDB specific UI in another place.

This PR adds one of those places, but it may very well be there are
other places that this should be added.

Fixes https://github.com/eclipse-cdt/cdt/issues/972