Remove generation of adoc files in code cleanliness
This was supposed to be a check to ensure adoc generation was
working, but it interferes with the overall check code cleanliness
as it leads to jgit dirty errors.

The generation of the docs is done as part of the build too, so
it is redundant to do them here.