Run DSF-GDB tests when releng may have changed (#1042)
On a recent update to the target platform, the DSF-GDB tests
did not run because none of the changes matched the dsf
filter. Going forward, include running DSF tests when any
releng changes happen.

This replaces commit 8d0642568c0. That commit ran the tests
all the time, which was too slow.

Part of #1037