Try automatically re-running failed tests
This will try to rerun failed tests up to three times.
The side effect is that a very broken test run will take much
longer to complete, but it is hoped for those few
https://github.com/eclipse-cdt/cdt/labels/flakytests in the system
this will resolve the issue.

Note this is applied only to the GitHub actions run, not to Jenkins
run where we build releases from. Ideally we don't want to do that
there.

In a future commit we can consider removing flakyTest from -DexcludedGroups
too.

See [surefire docs](https://maven.apache.org/surefire/maven-surefire-plugin/examples/rerun-failing-tests.html)
for details.