Add a directory tracker to the CBS Makefile builder
Messages such as "make: Entering directory" are now processed so
that the scanner info builder knows the correct full path to use
for build.

Without this code calls to ToolChain.getResourcesFromCommand() would
pass in the wrong directory and resources could not be reliably
calculated. With the incorrect calculation the scanner info does
not get applied to the file correctly and the file cannot be indexed
reliably.

TODO:

- [ ] Decide how to handle make -j as "make: Entering directory"
becomes unreliable
- [ ] Consider whether we should search project for resources
so that the exact build directory is less important
- [ ] Consider how to report back to the user that the detected
resource was not found in the project.