Include system environment when debugging core build
When doing a local launch, the only environment passed to the debuggee
was the modified variables, typically just PATH. This meant all the
other environment was lost.

This change does the same thing as all the other calls to setBuildEnvironment
and passes in the initial system environment.

I added a small note to the API of setBuildEnvironment to try
to avoid this error in the future.